pyproject.toml
[project]
name = "sendly"
version = "0.1.0"
description = ""
authors = [
    {name = "mustela",email = "ejankowski@gmail.com"}
]
packages = [
    { include = "app" }
]
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry]
packages = [
    { include = "app" }
]

[tool.poetry.scripts]
dev = "run:dev"
worker = "run_worker:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.115.11"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.39"
asyncpg = "^0.30.0"
gunicorn = "^23.0.0"
alembic = "^1.15.1"
faker = "^37.1.0"
psycopg2-binary = "^2.9.10"
pydantic = {extras = ["email"], version = "^2.11.3"}
black = "^25.1.0"
ruff = "^0.11.6"
pydantic-settings = "^2.9.1"
python-jose = "^3.3.0"
opentelemetry-instrumentation-fastapi = "^0.53b1"
opentelemetry-exporter-otlp-proto-grpc = "^1.32.1"
opentelemetry-sdk = "^1.32.1"
requests = "^2.32.3"
opentelemetry-instrumentation-sqlalchemy = "^0.53b1"
pyjwt = "^2.10.1"
sqlalchemy-json = "^0.7.0"
cryptography = "^44.0.3"
fernet = "^1.0.1"
celery = "^5.5.2"
rollbar = "^1.3.0"
fastmcp = "^2.6.0"
redis = ">=5.0.1,<6.0.0"
fastapi-pagination = "^0.14.0"
# tessera-sdk = { path = "../tessera-sdk-py/", develop = true }
tessera-sdk = { git = "https://github.com/tesserahq/tessera-sdk-py.git", branch = "main" }
python-slugify = "^8.0.4"
pytz = "^2025.2"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.28.0"  # Required by FastAPI TestClient
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
